==> Part1_Getting_Started.ipynb <==
<h1>Getting Started with Python</h1>
<h3>Overview</h3>
<p>
    <ul>
        <li>Overview of Python, Anaconda, and Jupyter Notebooks</li>
        <li>Data Values and interpretations</li>
        <li>Variables, Functions, and Arguments</li>
        <li>Comparisons, Conditional, and Loops</li>
        <li>Libraries</li>
    </ul>
</p>

<h3>Resources</h3>
<p>
    <ul>
        <li>O'Reilly Learning Platform: <a href = "https://databases.lib.wvu.edu/connect/1540334373" target ="_blank">https://databases.lib.wvu.edu/connect/1540334373</a></li>
    </ul>
</p>

<h3>Basic Concepts _ What is Python and Programming Language Software</h3>
<p>Python is a programming language software packages that allows you to give commands to your computer.</p>
<ul><li><a href ="https://www.python.org/">https://www.python.org/</a></li></ul>

<h3>Why use Python</h3>
<p>
    <ul>
        <li>Rich Ecosystem of Libraries</li>
        <li>Integration with Other Tools and Languages</li>
        <li>Platform Independent and Non-proprietary</li>
        <li>Reproducibility and Transparency</li>
        <li>Integrates into Proprietary Software</li>
        <li>Shareable</li>
        <li>Add-on Libraries</li>
   </ul>
</p>

<h3>Jupyter Notebooks</h3>
<p><a href ="https://jupyter.org/" target="_blank">https://jupyter.org/</a></p>
<p>Jupyter Notebooks are an interactive web-based tool that allows you to create and share documents that contain live code, equations, visualizations, and narrative text.</p>

<h3>Anaconda</h3>
<p><a href ="https://www.anaconda.com/" target="_blank">https://www.anaconda.com/</a></p>
<p>Anaconda is a popular and powerful distribution of Python and R programming languages specifically tailored for data science, machine learning, and scientific computing.</p>

<h3>Google Colab</h3>
<p><a href="https://colab.research.google.com/" target="_blank">https://colab.research.google.com/</a></p>
<p>a free cloud-based platform provided by Google that allows you to write, execute, and share Python code in a collaborative environment. </p>


<h2>Getting Started</h2>
<p>
    <ol>
        <li>Open Anaconda Navigator</li>
        <li>Open Jupyterlab</li>
        <li>Using the directory on the right, select the folder you will be working in</li>
        <li>Start a new notebook</li>
        <li>Select the python kernel</li>
        <li>Click on File > <strong>Save Notebook As</strong> and give the file a name</li>
        <li>Click on File > <strong>Save Current Workspace As</strong> and save the workspace</li>
    </ol>
</p>

<h3>Inputting and Running Code</h3>
<p>In Jupyter Notebooks you input your code into cells. 
    <ul>
        <li>You can add new cells by clicking on the plus buttons in the cell or in the top menu.</li>
        <li>You can change how the kernel is interpreting the cell. By default it is code.</li>
        <li>You can run the cell by clicking on the play button or using the <strong>keyboard shortcut: CTRL + Enter (PC) / CMD + Return (MAC)</strong>.</li>
    </ul>    
</p>


#lets try and input code and run a cell

2+2

<h3>Commenting</h3>
<p>Since use will be performing several operations in a single document and
even in a particular code chunk, it becomes important to document what
processes you were performing or make notes to use for yourself or
others about your intentions.</p>
<p>Entering a hastag (\#) into your code will comment anything that comes after for one single line.</p>

# what is the mean of the variable

import numpy as np # load numpy

example = [1,50, 100, 1000]

average = np.mean(example)

# Get the mean

print("The mean of this list is", average)


<h2>Data Values in Python</h2>
<ul>
    <li>String – string of characters with no numeric value – “hello world” “26501”</li>
    <li>Integer – whole number</li> 
    <li>Float – number w/ decimal place</li>
    <li>Boolean – t/f – True or False </li>
    <li>None – nothing, nul, nil </li>
</ul>

#Type functions

type("26501") #string

type(4) #integer

type(4.5673) #float

type(True) #boolean

#Strings - anything entered in "" will be interpreted as a string

hello = "Hello World"
type(hello)

<h2>Python Standard Library</h2>
<p><a href ="https://docs.python.org/3/library/index.html">https://docs.python.org/3/library/index.html</a></p>

#Help function

help(print)

<h3>Fundamental Concepts in Python</h3>
<ul>
    <li>Variables</li>
    <li>Types</li>
    <li>Functions</li>
    <li>Libraries</li>
    <li>Comparison</li>
    <li>Conditionals</li>
    <li>Looping</li>
    <li>Lists</li>
</ul>

<h2>Variables</h2>
<p>Variables are containers for storing data values.</p>
<ul>
    <li>Syntax: named_container = value_assigned</li>
    <li>x = 5</li>
    </ul>

#Examples 

x = 5
y = "Hello World"
z= [1, 2, 3]

<h3>Explore Variables</h3>

#list created variables

%whos

# call the variables

print(z)

# data values for variables

type(z)
print("Type of values of z:", type(z))

# length of the variable

len(z)

<h2>Functions</h2>
<p>Functions are how you give commands using python code.</p>
<p>It is highly suggested to use <strong>TAB</strong>, use the <strong>documentation</strong> for libraries, and use the <strong>help function</strong> to understand what functions and arguments are available to you.</p>

<h3>Built-In Functions</h3>

<p><a href = "https://docs.python.org/3/library/functions.html" target="_blank">Standard Library: https://docs.python.org/3/library/functions.html</a></p>
<p><a href = "https://pandas.pydata.org/docs/reference/general_functions.html" target="_blank">Pandas: https://pandas.pydata.org/docs/reference/general_functions.html</a></p>
<p>Built-in functions are pre-defined functions that are available as part of the core language. They are build into the standard library as well as any loaded libraries.</p>

<h3>Create a Function</h3>

#You can create functions using def

def my_function():
  print("Hello from a function")

my_function()

<h3>Exercise: Create a Function</h3>

# Step 1: Create Variables

name = input("What is your name: ")
age = input("What is your age: ")


#Step 2: Define the function

def demo(name, age):
    print(name, age)
   
#Step 3: Call the function

demo(name, age)

<h2>Indentation</h2>
<p>In Python, indents are used to define the structure of the code. Unlike some other programming languages that use curly braces {} or keywords to indicate blocks of code, Python uses indentation. This makes the code more readable and easier to understand.</p>

def check_number(number):  # Starts the function; the indented lines below are part of this function
    if number > 0:
        print("The number is positive.")
    elif number < 0:
        print("The number is negative.")
    else:
        print("The number is zero.")  # This is where the function's code ends

print("Hello World")  # This starts a new section of code


def check_number(number):
print("The number is positive.")  # This line will cause an IndentationError

<h2>Arguments</h2>
<p>Arguments are the values that are passed to a function when it is called. In Python, functions can accept zero or more arguments. Arguments are separted by <strong>commas ( , )</strong>.</p>

#example of a using arguments with the round function

round(3.141592653589793)
round(3.141592653589793, 4)

<h2>Comparisons and Boolean Operators</h2>
<p>Use comparison operators to determine if objects in python are identical to each other.</p>
<ul style ="font-size:1.15em;">
    <li>Equal ( == )</li>
    <li>Not equal ( != )</li>
    <li>Greater than ( > )</li>
    <li>Less than ( < )</li>
    <li>Greater than or equal ( >= )</li>
    <li>Less than or equal ( <= )</li>
    </ul>

#Comparison

fruit1= "apple"
fruit2= "orange"

#Are items/variables == to eachother
fruit1 == fruit2

#Comparison

fruit1= "apple"
fruit2= "apple"

#Are items/variables == to eachother
fruit1 == fruit2

#Comparison

fruit1= "apple"
fruit2= "Apple"

#Are items/variables == to eachother
fruit1 == fruit2

#Comparison

number1 = 1
number2 = 2

#Are items/variables == to eachother
number1 != number2

##boolean operators - > Can be combined variables/inputs with boolean operators AND, OR, NOT

6 > 3 and 1 < 2

#Can check existence of an object in a string

"n" in "mississippi"

<h2>Conditionals</h2>
<p>Python operators that looks to see if an object meets stated conditions and will then run operations based of those determinations.</p>
<ul>
    <li>If = If the conditions made in the statement are met then perform the operation.</li>
    <li>Else = If the conditions made in the statement are NOT met then perform another operations.</li>
    <li>Elif = If the condition made in the statement are NOT met by either the IF statement or the ELSE statement perform this operation.</li>
    </ul>

#Conditionals

hungry = True

if hungry:
    print("Go eat something")
    
print("Continue with your day")

#Conditionals

hungry = False

if hungry:
    print("Go eat something")
else:
    print("Eat this anyway")

#Multiple Criteria elif
#elif water_temp == 0: --> not bad

water_temp = 0

if water_temp < 0:
    print("brrrrrr")
elif water_temp > 100:
    print("tccchhhhh")
else:
    print("I can drink this")

<h2>Loops</h2>
<p>The requested operation will repeat until it is told to stop.</p>

<h3>For Loops</h3>
<p>The 'for' loop is typically used when you know the number of iterations in advance or when you want to iterate over a sequence or an iterable object.</p>
<p><em>syntax: <strong>for</strong> variable <strong>in</strong> iterable</em></p>
<p>
    <ul>
        <li>variable: This is a placeholder variable that will take on the value of each element in the iterable object during each iteration of the loop.</li>
        <li>
iterable: This is the object over which the loop iterates. It can be a sequence (like a list, tuple, or string) or any other iterable object.</li>
    </ul>
</p>

#for loop

fruits = ["apple", "banana", "cherry"]
for fruit in fruits:
    print(fruit)

<h3>While Loops</h3>
<p>The while loop is typically used when you don't know the number of iterations in advance or when you want to loop as long as a certain condition is true.</p>

while True:
    print("It's still true")

x = 0
while x <= 20:
    print(x)
    x = x + 1
    
print("Stop looping")

x = 0
while x <= 20:
    if x == 8:
        break
    print(x)
    x = x + 1

<h2>Libraries</h2>
<p>Libraries in Python are collections of pre-written code that provide ready-made functions and tools for common tasks. They save you time and effort by providing solutions to problems that programmers commonly encounter.</p>


import datetime

# Get the current date and time
current_date = datetime.datetime.now()

# Print the current date and time
print("Today's date and time:", current_date)


#view installed libraries

# !pip list
!conda list

#search for library

!conda search beautifulsoup

#update pacakge

!conda update beautifulsoup4

<h3>Call A Library</h3>
<p>To use the functions of a library you must call it during your current kernel.</p>

<h3>Import</h3>
<p>Using import for a library will bring all functions for the library into your workspace</p>

import math

# Now you can use any tool from the math toolbox
print(math.sqrt(16))  # Output: 4.0
print(math.pi)        # Output: 3.141592653589793


<h3>From</h3>
<p>Using from will just bring one particular function from the library into your workspace.</p>

from math import sqrt

# Now you can use just the sqrt tool directly
print(sqrt(16))  # Output: 4.0


<h3>Installing a Library</h3>

<p>You install Python Libraries using the terminal</p>
<p>
    <ul>
        <li>Go to File > New > Terminal</li>
        <li>Copy the following <strong>conda install -c conda-forge geopandas</strong> and past into the terminal.</li>
    </ul>
</p>



<h3>Important Libraries for Data Science</h3>
<ul>
    <li><a href="https://pandas.pydata.org/">Pandas</a></li>
    <li><a href="https://matplotlib.org/">Matplotlib</a></li>
    <li><a href="https://numpy.org/">NumPy</a></li>
    <li><a href="https://www.scipy.org/">SciPy</a></li>
    <li><a href="https://plotly.com/python/">Plotly</a></li>
    </ul>

<h3>Magic Commands</h3>
<p>Magic commands in Python, specifically in environments like Jupyter Notebook, are special commands that help you perform various tasks more easily.</p>

%time
