==> Part2_Working_with_Data.ipynb <==
<h1>Working with Data with Python</h1>
<h3>Overview</h3>
<ul>
    <li>Entering your data [lists, dictionaries, series dataframes]</li>
    <li>Reading datasets and cleaning data</li>
    <li>Analyzing and visualizing data</li>
       </ul>


<h3>Resources</h3>
<p>
    <ul>
        <li>O'Reilly Learning Platform: <a href = "https://databases.lib.wvu.edu/connect/1540334373" target ="_blank">https://databases.lib.wvu.edu/connect/1540334373</a></li>
    </ul>
</p>

<h2>Sequence Types (Lists, Tupples, Ranges)</h2>
<p>ordered sets that we can enter and retrieve information from.</p>
<ul><li><a href = "https://docs.python.org/3/library/stdtypes.html#sequence-types-list-tuple-range">Python Documentation - Sequence Types</a></li></ul>

<h3>Lists</h3>
<p>Lists are used to store multiple items in a single variable. Lists are one of 4 built-in data types in Python used to store collections of data, the other 3 are Tuple, Set, and Dictionary, all with different qualities and usage.</p>
<p>named_varible = ["item1", "item2", "item3"]</p>

#List integer

my_var = [1, 2, 3, 4]
print(my_var)
type(my_var)

#List string

cities = ["Morgantown", "Charleston", "Reedsville", "Huntington"] 

print(cities)
type(cities)

#Indexing
#method returns the position at the first occurrence of the specified value.
#syntax []

cities [0] #numbered order starts at 0
#force an error messages with print(cities[4])

#cities [0:2] #range

#Adding and removing items to a list .append and .remove

cities.append("Logan's Run")
cities

#remove
#cities.remove("Logan's Run")
#cities

#What is "in" your list

len(cities)

#"Morgantown" in cities

# Define a list of numbers
numbers = [1, 2, 3, 4, 5]

# Loop over each number in the list
for number in numbers:
    # Print each number
    print(number + 1)

<h2>Functions in Standard Python that are similiar to lists</h2>
<ul>
    <li>Tupples -> A tuple is a collection which is ordered and unchangeable. Tuples are written with round brackets().</li>
    <ul><li>( _ )</li>
    </ul>
    <li>Sets -> A set is a collection which is unordered, unchangeable, and unindexed. Sets are written with curly brackets.</li>
    <ul><li>{ _ }</li>
    </ul>
    <li>Dictionaries -> A dictionary is a collection which is ordered*, changeable and does not allow duplicates. Dictionaries are written with curly brackets, and have keys and values.</li>
    <ul><li>{key1: value1, key2: value2, key3:value3 }</li>
    </ul>
    </ul>
    

#dictionaires
#one-domensional key-value list

country_roads = {"Morgantown": 30847, "Charlestown": 45879, "Reedsville": 603, "Huntington": 44934}
print(country_roads)
#print(country_roads["Reedsville"])

#add to dictionary

country_roads["Westover"]  = 4220
country_roads

<h2>Data science libraries in Python</h2>
<p>Listed below are the major libraries that provide built-in functions, methods, and constants that are important for doing data science analysis. Each library has a website with documentation (remember the Python Standard Library) that is great for reference and tutorials.</p>
<h3>Storage, Manipulations, Calculations</h3>
<ul>
    <li><a href ="https://numpy.org/">Numpy</a></li>
    <li><a href ="https://pandas.pydata.org/">Pandas</a></li>
    <li><a href="https://www.scipy.org/scipylib/index.html">Scipy</a></li>
    <li><a href="https://www.statsmodels.org/stable/index.html">StatsModels</a></li>
    </ul>
 <h3>Vizualization</h3>
 <ul>
    <li><a href="https://matplotlib.org/">Matplotlib</a></li>
    <li><a href ="https://bokeh.org/">Bokeh</a></li>
    </ul>
<h3>Machine Learning</h3>
  <ul>
    <li><a href ="https://scikit-learn.org/stable/">SciKit</a></li>
    <li><a href="https://www.tensorflow.org/">TensorFlow</a></li>
    <li><a href = "https://keras.io/">Keras</a></li>
    </ul>

<h2>Pandas</h2>

#calling a library

import pandas as pd

#Alias -> 
#you can call libraries as aliases using "as". This will allow you to simplify your code and the amount of typing that you
#need to do.

<h3>Series in Pandas</h3>
<p>Like a dictionary in the standard library, a series allows you to store key-value pairs in python.</p>
<ul><li><a href ="https://pandas.pydata.org/docs/reference/api/pandas.Series.html">Pandas Documentation - Series</a></li></ul>

#series - A Pandas Series is like a column in a table. It is a one-dimensional array holding data of any type.

wvu_towns = pd.Series(country_roads)

wvu_towns

series_example = pd.Series([1, 2, 3, 4], index=['a', 'b', 'c', 'd'])
series_example

<h2>Data Frames</h2>
<p>Dataframes are 2 dimensional data structures containing key-value pairs.

import pandas as pd

# Creating individual Series for each column
year_series = pd.Series([1977, 1980, 1983], name="Year")
title_series = pd.Series(["Star Wars", "Empire Strikes Back", "Return of the Jedi"], name="Title")
length_series = pd.Series([121, 124, 144], name="Length")
gross_series = pd.Series([787, 534, 572], name="Gross")


# Displaying the DataFrame
starwars_df


# Combining the Series into a DataFrame

starwars_df = pd.DataFrame({
    "Year": year_series,
    "Title": title_series,
    "Length": length_series,
    "Gross": gross_series
})

#subsetting

starwars_df['Title']

# data values

starwars_df.dtypes

# amending data values

starwars_df['Year'] = starwars_df['Year'].astype(str)

starwars_df.dtypes

# retrieve descriptive statistics

# starwars_reviews.mean()

starwars_df["Length"].mean()


<h2>Reading Data</h2>

<h3> Read Options</h3>
<ul><li>read_csv -- Load delimited data from a file, URL, or file-like object; use comma as default delimiter</li>
<li>read_fwf -- Read data in fixed-width column format (i.e., no delimiters)</li>
<li>read_excel -- Read tabular data from an Excel XLS or XLSX file</li>
<li>read_html -- Read all tables found in the given HTML document</li>
<li>read_json -- Read data from a JSON (JavaScript Object Notation) string representation, file, URL, or file-like object</li>
<li>read_sas -- Read a SAS dataset stored in one of the SAS system’s custom storage formats</li>
<li>read_spss -- Read a data file created by SPSS</li>
<li>read_stata -- Read a dataset from Stata file format</li>
<li>read_xml -- Read a table of data from an XML file</li></ul>

# read spss file

import pandas as pd

demographics = pd.read_spss("demographics.sav")
demographics

# may need to install pyreadstat - pip install pyreadstat

# read a csv file

import pandas as pd

reviews = pd.read_csv("customer_reviews.csv")
%whos

<h3>Exploring the Dataframe</h3>

#Viewing dataframes .head() . tail()

reviews.head() #default is 5

#reviews.tail(10) #change the number displayed

#viewing data values in dataframe

reviews.dtypes

#look as basic information about the dataframe with info()

reviews.info()

#viewing observations and variables numbers

reviews.shape

#viewing the names of variables

reviews.columns

# get basic descriptive statistics

reviews.describe()

#reviews.mean()
#reviews.median()
#reviews.mode()



<table style = "width:60%; margin-left:0; border: 3px solid #f0f0f0;">
  <tr style = "text-align: left;">
    <th>Aggregation</th>
    <th>Returns</th>
  </tr>
  <tr>
    <td>count</td>
    <td>Total number of items</td>
  </tr>    
  <tr>
    <td>first, last</td>
    <td>First and last item</td>
  </tr>   	
  <tr>
    <td>mean, median</td>
    <td>Mean and median</td>
  </tr>  
  <tr>
    <td>min, max</td>
    <td>Minimum and maximum</td>
  </tr>  
  <tr>
    <td>std, var</td>
    <td>Standard deviation and variance</td>
  </tr>  
  <tr>
    <td>mad</td>
    <td>Mean absolute deviation</td>
  </tr>  
  <tr>
    <td>prod</td>
    <td>Product of all items</td>
  </tr>  
  <tr>
    <td>sum</td>
    <td>Sum of all items</td>
  </tr> 
</table>

# viewing the values counts of observations in a variable

reviews["Class_Name"].value_counts()

<h2>Cleaning Data</h2>

<h3>Why would you need to clean data</h3>
<ul>
    <li>Data in columns and rows are not ordered in the correct way</li>
    <li>Creating values or ignoring missing data</li>
    <li>Units are not correct or are wrong in some way</li>
    <li>Order of magnitude is off</li>
    <li>Outliers and skewing of the data</li>
    </ul>

<h3>Missing Values</h3>

#check for missing data

missing_values = reviews.isna()
missing_values



# number of missing data

number_missing = reviews.isna().sum()
number_missing

#Get a report

# Check for missing values
print("Missing values (True indicates missing):")
print(reviews.isna())

# Count missing values in each column
print("\nCount of missing values per column:")
print(reviews.isna().sum())

# Get a summary of the DataFrame
print("\nDataFrame info:")
reviews.info()

# Check if there are any missing values in the entire DataFrame
print("\nAre there any missing values in the entire DataFrame?")
print(reviews.isna().any().any())

# Check if there are any missing values in each column
print("\nAre there any missing values in each column?")
print(reviews.isna().any())

# Check for non-missing values
print("\nNon-missing values (True indicates non-missing):")
print(reviews.notna())

# remove missing values

reviews.shape

#remove all observations with na

reviews_na = reviews.dropna()
reviews_na.shape

# remove 

# replace missing values

#replace na values

reviews["Title"] = reviews["Title"].fillna("None Given")
reviews["Title"]

#get value counts for a variable

reviews["Title"].value_counts()

<h3>Change Data Values</h3>

#view the dataframe

reviews

# view types of data values in the dataframe

reviews.dtypes

# change the values 

# Clothing ID should be a string
reviews['Clothing_ID'] = reviews['Clothing_ID'].astype("str")

# Recommended_IND, Division_Name, Department_Name, and Class_Name should be a category
reviews[['Recommended_IND', 'Division_Name', 'Department_Name', 'Class_Name']] = reviews[['Recommended_IND', 'Division_Name', 'Department_Name', 'Class_Name']].astype("category")

# view the values

reviews.dtypes


reviews["Division_Name"].value_counts()

<h3>Datetime Values</h3>

import pandas as pd

# load scotus approval dataframe

scotus = pd.read_csv("scotus_approval.csv")
scotus


scotus["date"] = pd.to_datetime(scotus["end_date"])
scotus.dtypes

<h3>Boolean Operators</h3>
<p>Use comparison operators to determine to filter observations in a variable.</p>
<ul style>
    <li>Equal ( == )</li>
    <li>Not equal ( != )</li>
    <li>Greater than ( > )</li>
    <li>Less than ( < )</li>
    <li>Greater than or equal ( >= )</li>
    <li>Less than or equal ( <= )</li>
    </ul>

<h3>Filter Variables</h3>
<p>Filter out one or more observations from a variable using boolean operators to set criterias.</p>

# get mean of ratings

reviews["Rating"].mean()

#filter observations using boolean operators

#find the mean of rating that people who bought from the General Department 

reviews_filter = reviews[reviews["Division_Name"]== "General"]
reviews_filter["Division_Name"].value_counts()

reviews_filter["Rating"].mean()

reviews["Class_Name"].value_counts()

#filter observations in mulitple variables

reviews_filter_2 = reviews[(reviews["Division_Name"] == "General") & (reviews["Class_Name"] != "Sweaters")]
reviews_filter_2["Rating"].mean()

<h3>Select Variables</h3>
<p>Select or remove variables from the dataframe</p>

scotus.columns

# select columns from the dataframe

scotus = scotus[['poll_id',  'pollster', 'yes', 'no', 'alternate_answers', 'date']]
scotus

scotus.columns

# removed columns from the dataframe

scotus = scotus.drop(columns = ['alternate_answers'])
scotus

<h3>Sample</h3>
<p>Filter random selections from the dataframe</p>

# number

reviews_sample = reviews.sample(n=500)
reviews_sample.describe()

# fraction

reviews_sample = reviews.sample(frac=0.25, replace = True, random_state=1)
reviews_sample.describe()

<h3>Assign</h3>
Create new columns or modify existing ones in a dataframe. 

demographics

# lets create a new column that divides the income variable by 1000

demographics = pd.read_spss("demographics.sav")

demographics = demographics.assign(income = (demographics["income"]/1000))
demographics


<h3>Recode Variables</h3>
<p>Transform the values of a variable into new values based on specific criteria</p>

reviews["Recommended_IND"].value_counts()

#overwrite the variable

reviews_recode = reviews["Recommended_IND"].replace([1, 0], ["Yes", "No"])
reviews_recode

# create a new variable

reviews = reviews.assign(Recommended_recode = reviews["Recommended_IND"].map({0: "No", 1: "Yes"}))
reviews

# assign creates and new variable and # map sets the new values

reviews["Recommended_recode"].value_counts()

<h3>Rename Variables</h3>
<p>Rename the column</p>

reviews = reviews.rename(columns={"Recommended_IND": "Recommended_num", "Recommended_recode":"Recommended_label"})
reviews

<h3>Relocate Variables</h3>
<p>Move the column location in the dataframe</p>

reviews.columns

new_order = ['Clothing_ID', 'Age', 'Title', 
             'Review_Text', 'Rating', 'Recommended_num', 'Recommended_label',
             'Positive_Feedback_Count', 'Division_Name', 
             'Department_Name', 'Class_Name']

reviews = reviews[new_order]
reviews

<h3>Sort Variables</h3>
<p>Rearrange the observations in a column</p>

reviews.sort_values(by="Rating", ascending=False)

<h3>Bin Observations</h3>
<p>Group observations into 'bins' or categories based on their values. </p>

#bin values

reviews = reviews.assign(Age_Category = lambda x: pd.cut(x['Age'], bins = [18, 35, 65, float('inf') ], labels=['Young', 'Middle-aged', 'Old']))
reviews["Age_Category"].value_counts()

# We use the assign method to add a new column "Age_Category" to the DataFrame.
# The lambda function passed to assign calculates the age category based on the "Age" column using pd.cut.
# The pd.cut function bins the ages into categories defined by the specified bins and assigns corresponding labels.

<h3>Pivot Table</h3>
<p>Allows you to summarize and analyze the data by aggregating values across different dimensions, such as rows and columns,</p>

reviews.dtypes

reviews_pivot = reviews.pivot_table(
    values = "Rating",
    index = "Recommended_label",
    columns="Class_Name",
    aggfunc = "mean",
    fill_value = 0
)

reviews_pivot


<h3>Groupby</h3>
<p>Groups your observations in one or more varaiables.</p>

#Groupby

class_rating = reviews[["Rating", "Class_Name"]].groupby("Class_Name", observed=True).mean()
class_rating


#class_rating = reviews[["Rating", "Class_Name"]].groupby("Class_Name", observed=True).describe()
#class_rating


<h3>Aggregate</h3>
<p>Combine multiple data values into a single summary statistic. For example, finding the sum, mean, median, minimum, maximum, or standard deviation of a group </p>

result = reviews.groupby('Recommended_label', observed=True)[['Rating', 'Positive_Feedback_Count']].agg(['min', 'mean', 'std', 'max'])
result

<h3>Dummy Variables</h3>
<p>Turns categories into numbers so we can use them in analyses.</p>

reviews_dummies = pd.get_dummies(reviews["Division_Name"], prefix = "Division")

reviews_dummies

# Convert to 1 and 0

reviews_dummies = reviews_dummies.astype(int)
reviews_dummies.mean()

<h3>Concat Dataframes</h3>
<p>Append dataframes with matching observations.</p>

reviews_combined = pd.concat([reviews, reviews_dummies], axis=1)
reviews_combined

<h3>Model Building with Dummies</h3>
<p>
    <ul>
        <li>Use categorical data in machine learning and statistical models.</li>
        <li>Avoid implying ordinal relationships in categorical data.</li>
        <li>Capture categorical information accurately.</li>
        <li>Enhance flexibility in model building and analysis.</li>
    </ul>
</p>

reviews_combined.columns

from sklearn.linear_model import LinearRegression

# Define the features (dummy variables) and target (Sales)
X = reviews_combined[['Division_General', 'Division_General Petite', 'Division_Initmates']]
y = reviews_combined["Rating"]

# Create and fit the model
model = LinearRegression()
model.fit(X, y)

# Print the coefficients
print("\nCoefficients of the regression model:")
print(model.coef_)
print("Intercept:", model.intercept_)

<h3>Export Dataframes</h3>

#Write

reviews.to_csv("cleaned_reviews.csv")
reviews_pivot.to_csv("pivot_reviews.csv")
class_rating.to_csv("class_ratings.csv")
scotus.to_csv("scotus.csv")
